requirements.txt
alabaster==0.7.13
asgiref==3.6.0
Babel==2.12.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-address==0.2.8
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.11
platformdirs==3.5.0
Pygments==2.15.1
PyJWT==2.7.0
pytz==2023.3
requests==2.30.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
